update rutas y library
Se ajusat configuracion de backend library ya que el dashboard de user estaba la logica que deberia estar en library
