+++ boo-books/src/components/searchbook.js
@@ -1,36 +1,24 @@
-import React from "react";
+import React from 'react';
 
-const Books = ({ books, searchTerm }) => {
+const Books = ({ books, takeLibro, returnLibro }) => {
   return (
-    <div style={{ display: "flex", flexDirection: "row" }}>
-      <div style={{ flex: 1 }}>
-        {books
-          .filter((book) =>
-            book.nombre.toLowerCase().includes(searchTerm.toLowerCase())
-          )
-          .map((book) => (
-            <div key={book._id}>
-              <img src={book.imagen} alt={book.nombre} style={{ width: "100px" }} />
-            </div>
-          ))}
-      </div>
-      <div style={{ flex: 1 }}>
-        {books
-          .filter((book) =>
-            book.nombre.toLowerCase().includes(searchTerm.toLowerCase())
-          )
-          .map((book) => (
-            <div key={book._id}>
-              <h3>{book.nombre}</h3>
-              <p>Autor: {book.autor}</p>
-              <p>Año de publicación: {book.publicacion}</p>
-              <p>ISBN: {book.isbn}</p>
-              <p>
-                Estado: {book.cantidadDisponible > 0 ? "Disponible" : "No disponible"}
-              </p>
-            </div>
-          ))}
-      </div>
+    <div>
+      {books.map(book => (
+        <div key={book._id} style={{ marginBottom: '10px', padding: '10px', border: '1px solid #ccc' }}>
+          <h3>{book.nombre}</h3>
+          <img src={book.imagen} alt={book.nombre} style={{ width: '150px', height: '200px' }} />
+          <p>Autor: {book.autor}</p>
+          <p>Género: {book.genero}</p>
+          <p>Cantidad Disponible: {book.cantidadDisponible}</p>
+
+          {book.cantidadDisponible > 0 ? (
+            <button onClick={() => takeLibro(book._id)}>Tomar Libro</button>
+          ) : (
+            <p style={{ color: 'red' }}>No disponible</p>
+          )}
+          <button onClick={() => returnLibro(book._id)}>Devolver Libro</button>
+        </div>
+      ))}
     </div>
   );
 };
